apache-cassandra/requirements.txt
blinker==1.8.2
cassandra-driver==3.29.1
click==8.1.7
Flask==3.0.3
geomet==0.2.1.post1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
six==1.16.0
Werkzeug==3.0.3
